# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Student Behavioral Wellness Tracker API
  version: 1.0.0
  description: REST API for student wellness tracking, alerts, messaging, and AI support.
servers:
  - url: http://localhost:5000
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    StudentCreateRequest:
      type: object
      required: [name, email]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        grade:
          type: string
        risk:
          type: string
          enum: [Low, Medium, High, Stable]
        temporaryPassword:
          type: string
          minLength: 8
    StudentUpdateRequest:
      type: object
      properties:
        mentor:
          type: string
        risk:
          type: string
          enum: [Low, Medium, High, Stable]
        status:
          type: string
          enum: [Happy, Calm, Stressed, Tired, Sad, Anxious, Excited]
        enrollmentStatus:
          type: string
          enum: [Active, Inactive]
        progressNotes:
          type: string
        historyEvent:
          type: object
          properties:
            type:
              type: string
            label:
              type: string
            details:
              type: string
...
    MessageCreateRequest:
      type: object
      required: [text]
      properties:
        studentId:
          type: string
        studentName:
          type: string
        text:
          type: string
    AiChatRequest:
      type: object
      required: [message]
      properties:
        message:
          type: string
        moodContext:
          type: string
        tone:
          type: string
          enum: [Encouraging, Humorous, Stoic]
paths:
...
  /api/students:
    get:
      tags: [Students]
      summary: List all students
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Student list
    post:
      tags: [Students]
      summary: Create a new student
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentCreateRequest'
      responses:
        '201':
          description: Student created
  /api/students/{id}:
    put:
      tags: [Students]
      summary: Update a student
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentUpdateRequest'
      responses:
        '200':
          description: Student updated
    delete:
      tags: [Students]
      summary: Delete a student and related records
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Student deleted
  /api/history:
    get:
      tags: [History]
      summary: Get combined student history
      security:
        - bearerAuth: []
      responses:
        '200':
          description: History list
...
  /api/alerts:
    get:
      tags: [Alerts]
      summary: List alerts
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Alert list
    delete:
      tags: [Alerts]
      summary: Clear all alerts
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Alerts cleared
...
  /api/messages:
    post:
      tags: [Messages]
      summary: Send a message
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageCreateRequest'
      responses:
        '201':
          description: Message created
...
  /api/ai-chat:
    post:
      tags: [AI]
      summary: Chat with WellnessBuddy
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AiChatRequest'
      responses:
        '200':
          description: AI reply returned